Add placeholder execute methods to all commands
