+++ src/main/java/seedu/lifeasier/commands/DisplayScheduleCommand.java
@@ -1,5 +1,8 @@
 package seedu.lifeasier.commands;
 
+import seedu.lifeasier.Ui;
+import seedu.notes.NoteList;
+
 public class DisplayScheduleCommand extends Command {
 
     private boolean isDisplayWeek;
@@ -8,4 +11,8 @@ public class DisplayScheduleCommand extends Command {
         this.isDisplayWeek = toDisplay.equals("week");
     }
 
+    @Override
+    public void execute(Ui ui, NoteList notes) {
+
+    }
 }
