+++ src/main/java/seedu/lifeasier/commands/ExitCommand.java
@@ -1,8 +1,16 @@
 package seedu.lifeasier.commands;
 
+import seedu.lifeasier.Ui;
+import seedu.notes.NoteList;
+
 public class ExitCommand extends Command {
 
     public ExitCommand() {
         this.isTerminated = true;
+    }
+
+    @Override
+    public void execute(Ui ui, NoteList notes) {
+
     }
 }
