+++ src/main/java/seedu/lifeasier/commands/HelpCommand.java
@@ -1,4 +1,12 @@
 package seedu.lifeasier.commands;
 
+import seedu.lifeasier.Ui;
+import seedu.notes.NoteList;
+
 public class HelpCommand extends Command {
+
+    @Override
+    public void execute(Ui ui, NoteList notes) {
+
+    }
 }
